+++ packages/preview-middleware-client/src/adp/quick-actions/fe-v4/utils.ts
@@ -1,0 +1,53 @@
+import { getControlById } from '../../../utils/core';
+import FlexCommand from 'sap/ui/rta/command/FlexCommand';
+import { QuickActionContext } from '../../../cpe/quick-actions/quick-action-definition';
+import CommandFactory from 'sap/ui/rta/command/CommandFactory';
+import { getAppComponent, getPageName, getReference } from '../../../utils/fe-v4/utils';
+
+export async function executeToggleAction(
+    context: QuickActionContext,
+    isButtonEnabled: boolean,
+    controlType: string,
+    propertyPath: string
+): Promise<FlexCommand[]> {
+    const controls = context.controlIndex[controlType] ?? [];
+    const control = controls[0];
+    if (control) {
+        const modifiedControl = getControlById(control.controlId);
+        if (!modifiedControl) {
+            return [];
+        }
+
+        const { flexSettings } = context;
+        const parent = modifiedControl.getParent();
+        if (!parent) {
+            return [];
+        }
+
+        const modifiedValue = {
+            reference: getReference(modifiedControl),
+            appComponent: getAppComponent(modifiedControl),
+            changeType: 'appdescr_fe_changePageConfiguration',
+            parameters: {
+                page: getPageName(parent),
+                entityPropertyChange: {
+                    propertyPath: propertyPath,
+                    propertyValue: !isButtonEnabled,
+                    operation: 'UPSERT'
+                }
+            }
+        };
+
+        const command = await CommandFactory.getCommandFor<FlexCommand>(
+            modifiedControl,
+            'appDescriptor',
+            modifiedValue,
+            null,
+            flexSettings
+        );
+
+        return [command];
+    }
+
+    return [];
+}